+++ CHANGES.md
@@ -1,6 +1,13 @@
 twilio-csharp Changelog
 =======================
 
+[2026-01-07] Version 7.14.1
+---------------------------
+**Api**
+- Added optional parameter `clientNotificationUrl` for create call api
+- Added optional parameter `clientNotificationUrl` for create participant api
+
+
 [2025-12-17] Version 7.14.0
 ---------------------------
 **Trunking**